refactor: 개별 패키지로 도메인 이동
