+++ src/main/java/kitchenpos/order/repository/OrderRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.repository;
+package kitchenpos.order.repository;
 
 import java.util.List;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
