+++ src/main/java/kitchenpos/ordertable/application/TableService.java
@@ -1,12 +1,12 @@
-package kitchenpos.application;
+package kitchenpos.ordertable.application;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.repository.OrderRepository;
-import kitchenpos.repository.OrderTableRepository;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.order.repository.OrderRepository;
+import kitchenpos.ordertable.repository.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
