+++ src/main/java/kitchenpos/ordertable/repository/OrderTableRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.repository;
+package kitchenpos.ordertable.repository;
 
 import java.util.List;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
