+++ src/main/java/kitchenpos/ordertable/ui/TableRestController.java
@@ -1,9 +1,9 @@
-package kitchenpos.ui;
+package kitchenpos.ordertable.ui;
 
 import java.net.URI;
 import java.util.List;
-import kitchenpos.application.TableService;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.ordertable.application.TableService;
+import kitchenpos.ordertable.domain.OrderTable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
